+++ src/Library.h
@@ -6,8 +6,8 @@
 
 namespace Lib {
 std::vector<float> generateRandomValues(int amount, float min, float max);
-std::vector<int> createGridPolygons(const std::vector<int>& vertices,
-                                    int gridWidth);
+std::vector<uint16_t> createGridPolygons(const std::vector<int>& vertices,
+                                         int gridWidth);
 
 double lowFrequencyOscillator(double frequency);
 glm::vec2 smoothstep(const glm::vec2& xy);
@@ -17,4 +17,4 @@ std::string upperToLowerCase(std::string string);
 // Cross platform functions
 std::string path(const std::string& linuxPath);
 std::string ifShaderCompile(std::string linuxPath);
-}
+}  // namespace Lib
